Introduced the concept of NodeOptionWirings and added a JSON editor for dictionaries in the editor. Closes #19
NodeOptionWirings allow the creator of a node to descibe relationships between option fields.

This has been implemented in the MessageFromConstDict node where the “dict” option and the “message_type” option now have a NodeOptionWiring. This allows the editor to update the dicts data based on the message type selected.

Additionally to ease the editing of dictionaries in the editor a jsoneditor was added that provides syntax highlighting and validation.