Fix bug where input and output bridge pub/sub were unregistered before being used.
Switch from actionlib actions to async services.

Async services are simpler and more reliable to use than actions, thus I switched to allow for parallel execution of nodes, etc.

Fix bug where executing a capability in a capability caused it to revert to the UNASSIGNED state.

Changed the Behavior to use the same debug manager for all local capabilities.